requirements.txt
streamlit>=1.32.0
pillow>=10.0.0
openai>=1.0.0
numpy>=1.24.0
requests>=2.31.0
